+++ NEWS.md
@@ -1,3 +1,20 @@
+# simmer 3.6.5
+
+## New features:
+
+* `set_attribute()` (and `set_global()` by extension) can set multiple attributes at once by providing vectors of `keys` and `values` (or functions returning such `keys` and/or `values`). `get_attribute()` (and `get_global()` by extension) can retrieve multiple `keys` (#122).
+* New `stepn()` method deprecates `onestep()` (e452975).
+
+## Minor changes and fixes:
+
+* Restore `ostream` after formatting (9ff11f8).
+* Fix arrival cloning to copy attributes over to the clone (#118).
+* Fix self-induced preemption through `set_capacity()` (#125).
+* Update "Queueing Systems" vignette (a0409a0, 8f03f4f).
+* Update "Advanced Trajectory Usage" vignette (4501927).
+* Fix print methods to return the object invisibly (#128).
+* New `Dining Philosophers Problem` vignette (ff6137e).
+
 # simmer 3.6.4
 
 ## Minor changes and fixes: